pyproject.toml
[project]
name = "pricetracker"
version = "0.1.0"
description = ""
authors = [
    {name = "CarlosMarrero98",email = "alu0101210995@ull.edu.es"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-telegram-bot[job-queue] (>=22.0,<23.0)",
    "matplotlib (>=3.10.3,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pdoc = "^15.0.3"

[tool.pytest.ini_options]
pythonpath = [
    "."
]